Ctrl+C para matar al proceso que se esta ejecutando

+++ x64BareBones-master/Userland/SampleCodeModule/processes/clear.c
@@ -1,4 +1,7 @@
-void clear(int argc, char **argv){
+#include "../include/userlib.h"
+#include "../include/syscall.h"
+
+void clear(uint64_t argc, char **argv){
     clearScreen();
     my_exit();
 }
