+++ x64BareBones-master/Userland/SampleCodeModule/processes/registers.c
@@ -1,4 +1,7 @@
-void registers(int argc, char **argv){
+#include "../include/userlib.h"
+#include "../include/syscall.h"
+
+void registers(uint64_t argc, char **argv){
     print_registers();
     my_exit();
 }
