+++ x64BareBones-master/Userland/SampleCodeModule/processes/time.c
@@ -1,4 +1,7 @@
-void time(int argc, char **argv){
+#include "../include/userlib.h"
+#include "../include/syscall.h"
+
+void time(uint64_t argc, char **argv){
     printTime();
     my_exit();
 }
